requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
databases[sqlite]
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
rich